feat : create connection to mysql, utilize flyway for database migration,...

See merge request I528859/huister!33

+++ src/main/java/nl/fontys/s3/huister/business/interfaces/user/GetUserByIdUseCase.java
@@ -4,5 +4,5 @@ package nl.fontys.s3.huister.business.interfaces.user;
 import nl.fontys.s3.huister.business.response.user.GetUserByIdResponse;
 
 public interface GetUserByIdUseCase {
-    GetUserByIdResponse getUserById(final int id);
+    GetUserByIdResponse getUserById(final long id);
 }
